Restructure CI manifests, document CI Jobs
Move CI manifests around to clearly separate resources:
- trigger template for all repos
- trigger template for plumbing
- task, ci jobs and conditions

Define a fixed set of parameters that trigger templates
may rely on from the event listener + interceptors +
overlays + bindings, even if this means certain params
may not be used by some CI jobs.

Change the PR kind label check to align to other checks
so move it from a TaskRun to a PipelineRun and move the
pipeline definition to its own manifest.

(TBD) Document the trigger template input interface
and the process for adding a new CI job.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>